+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the sdc11073 module will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Fixed
+
+- nsmap is not fully copied when copying report node in client mdib [#191](https://github.com/Draegerwerk/sdc11073/issues/191)
+
 ## [1.1.27] - 2023-06-29
 
 ### Added